# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOA Prototype API",
    "version": "0.1.0",
    "description": "Demo authentication and payments only. Not a production PG or customs API."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/rooms/{id}/replies": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": { "summary": "참여자 전용 상황별 기본 답장 추천 · 외부 AI 호출 없음", "responses": { "200": { "description": "기본 추천", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatReplies" } } } }, "403": { "description": "대화방 참여자 아님" } } },
      "post": { "summary": "동의 후 AI 초안 추천 · 미설정/실패 시 기본 추천", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["useAI"], "properties": { "useAI": { "type": "boolean", "enum": [true] } } } } } }, "responses": { "201": { "description": "AI 또는 기본 초안 · 자동 전송하지 않음", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatReplies" } } } }, "403": { "description": "대화방 참여자 아님" } } }
    },
    "/rooms/{id}/messages": {
      "post": {
        "summary": "참여자 채팅 메시지",
        "operationId": "post_rooms_id_messages",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageInput"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, random 24-hour demo session token"
      }
    },
    "schemas": {
...
      "MessageInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": [
          "text"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "ChatReplies": { "type": "object", "required": ["source", "suggestions", "aiAvailable", "notice"], "properties": { "source": { "type": "string", "enum": ["AI", "BASIC"] }, "suggestions": { "type": "array", "minItems": 1, "maxItems": 3, "items": { "type": "string", "maxLength": 100 } }, "aiAvailable": { "type": "boolean" }, "notice": { "type": "string" } } },
...
    }
  }
}